update code, the model can be trained now

+++ text_cnn.py
@@ -24,7 +24,7 @@ class TextCNN(object):
             # self.W = tf.Variable(
             #     tf.random_uniform([vocab_size, embedding_size], -1.0, 1.0),
             #     name="W")
-            self.W = tf.Variable(init_embedding, trainable=False, name="W")
+            self.W = tf.Variable(init_embedding, trainable=False, name="W", dtype='float32')
             self.embedded_chars = tf.nn.embedding_lookup(self.W, self.input_x)
             self.embedded_chars_expanded = tf.expand_dims(self.embedded_chars, -1)
